drafts/daydream-field-intel/requirements.txt
groq>=1.0.0
openai>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
duckduckgo-search>=6.0.0
rich>=13.0.0
click>=8.0.0
python-dotenv>=1.0.0
